feat: api-handlers-decomposition-recovery-007 - Split oversized server tests when file-count gate requires
