+++ pkg/api/server_test_helpers_test.go
@@ -478,3 +478,56 @@ func decodeJSONResponse[T any](t *testing.T, rec *httptest.ResponseRecorder) T {
 func encodeNextToken(token string) string {
 	return base64.StdEncoding.EncodeToString([]byte(token))
 }
+
+func decodeListWorkPage(t *testing.T, srv *Server, path string) factoryapi.ListWorkResponse {
+	t.Helper()
+
+	req := httptest.NewRequest(http.MethodGet, path, nil)
+	rec := httptest.NewRecorder()
+	srv.Handler().ServeHTTP(rec, req)
+	if rec.Code != http.StatusOK {
+		t.Fatalf("%s status = %d, want 200: %s", path, rec.Code, rec.Body.String())
+	}
+	return decodeJSONResponse[factoryapi.ListWorkResponse](t, rec)
+}
+
+func assertListedWorkIDs(t *testing.T, works []factoryapi.Work, want []string) {
+	t.Helper()
+	if len(works) != len(want) {
+		t.Fatalf("results = %d, want %d: %#v", len(works), len(want), works)
+	}
+	for i, wantWorkID := range want {
+		if got := stringValue(works[i].WorkId); got != wantWorkID {
+			t.Fatalf("result[%d].workId = %q, want %q: %#v", i, got, wantWorkID, works)
+		}
+	}
+}
+
+type upsertValidationFailureCase struct {
+	name    string
+	path    string
+	body    string
+	factory *testutil.MockFactory
+	wantMsg string
+}
+
+func runUpsertValidationFailureCases(t *testing.T, cases []upsertValidationFailureCase) {
+	t.Helper()
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			mf := tc.factory
+			if mf == nil {
+				mf = &testutil.MockFactory{}
+			}
+			mf.Marking = &petri.MarkingSnapshot{Tokens: make(map[string]*interfaces.Token)}
+			srv := newTestServer(mf)
+
+			rec := upsertWorkRequest(t, srv, tc.path, tc.body)
+			assertJSONError(t, rec, http.StatusBadRequest, "BAD_REQUEST", tc.wantMsg)
+			if len(mf.Submitted) != 0 {
+				t.Fatalf("submitted count = %d, want 0", len(mf.Submitted))
+			}
+		})
+	}
+}
